*+_MaterialTextFieldMoney: Con BigDecimal

+++ src/main/java/com/jhw/swing/material/components/textfield/validated/_MaterialTextFieldDouble.java
@@ -17,7 +17,7 @@ public class _MaterialTextFieldDouble extends _MaterialTextField<Integer> {
     }
 
     public _MaterialTextFieldDouble(boolean negative) {
-        super(Integer.class);
+        super(Double.class);
         this.negative = negative;
         addListeners();
     }
